Grade Calculator
This program calculates a student's current grade in my Introduction to Python class based on the number of unit deliverables, labs, and engagement points that have been completed at a satisfactory level. This lab demonstrates my understanding of branching structures. I left in my instructor's instructions for clarity's sake, but the code is my own.