Update fair_wage.py
This program detects whether or not you make a fair hourly wage based on the 30% rule, and then calculates what a fair wage would be if not. It is the first unit project for my Introduction to Python course, meant to test my understanding of variables, inputs, formatting, functions, and decision structures. All code was written by me.